refactor(G2): extract completion pipeline from Indexer into features/completion.rs
Move ~200 lines out of src/indexer.rs into src/features/completion.rs:

- run_completions(&Indexer, ...) contains the full pipeline (cache, lambda
  recv type resolution, dot-completion, bare-word, lambda-param completions)
- line_for_position, resolve_lambda_recv_type, add_lambda_param_completions
  are now module-private free functions in features/completion.rs
- before_cursor, split_prefix, dot_receiver (pure string helpers) moved
  to features/completion.rs; their tests moved to features/completion_tests.rs
- Indexer::completions() becomes a 1-line delegate to run_completions()
- No trait API changes; all inherent method call sites unchanged
- Gate complete_symbol, complete_symbol_with_context, is_annotation_context
  re-exports in resolver/mod.rs under #[cfg(test)] (only used by test files)

836 tests pass, zero new clippy warnings.
